Mise à jour des tests unitaires suite au refactoring
